fix retrieving stats

+++ internal/models/stats.go
@@ -3,8 +3,8 @@ package models
 import "time"
 
 type Statistics struct {
-	TotalProccessedMessages  	int
+	TotalProccessedMessages  	int64
 	LastProcessedMessageTime 	*time.Time
 	AverageProcessingTimeMs  	*float64
-	QueuedMessages				int
+	QueuedMessages				int64
 }
